requirements.txt
pytest
flake8
attrs
